Use std::size_t for register id

+++ simulator/func_sim/operation.h
@@ -160,7 +160,7 @@ public:
     using RegisterUInt = T;
     using RegisterSInt = sign_t<RegisterUInt>;
 
-    void set_v_src( const T& value, uint8 index)
+    void set_v_src( const T& value, size_t index)
     {
         if ( index == 0)
             v_src1 = value;
@@ -228,7 +228,7 @@ public:
     using MyDatapath = Datapath<T>;
     using Register = R;
     using RegisterUInt = T;
-    R get_src_num( uint8 index) const { return ( index == 0) ? src1 : src2; }
+    R get_src_num( size_t index) const { return ( index == 0) ? src1 : src2; }
     R get_dst_num()  const { return dst;  }
     R get_dst2_num() const { return dst2; }
